# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/adverts": {
            "get": {
                "description": "Fetch a list of all adverts with optional pagination.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adverts"
                ],
                "summary": "Retrieve all adverts",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit the number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of adverts",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PreviewAdvertCard"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid limit or offset",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve adverts",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new advert to the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adverts"
                ],
                "summary": "Create a new advert",
                "parameters": [
                    {
                        "description": "Advert data",
                        "name": "advert",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdvertRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Advert created",
                        "schema": {
                            "$ref": "#/definitions/dto.Advert"
                        }
                    },
                    "400": {
                        "description": "Invalid advert data",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create advert",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/adverts/cart/{cartId}": {
            "get": {
                "description": "Fetch a list of adverts in the specified cart.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adverts"
                ],
                "summary": "Retrieve adverts by cart ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cart ID",
                        "name": "cartId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of adverts in cart",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PreviewAdvertCard"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid cart ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve adverts by cart ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/adverts/category/{categoryId}": {
            "get": {
                "description": "Fetch a list of adverts associated with a specific category ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adverts"
                ],
                "summary": "Retrieve adverts by category ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "categoryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of adverts by category ID",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PreviewAdvertCard"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid category ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve adverts by category ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/adverts/my": {
            "get": {
                "description": "Fetch a list of adverts associated with a specific user ID.",
                "tags": [
                    "adverts"
                ],
                "summary": "Retrieve adverts by user ID",
                "responses": {
                    "200": {
                        "description": "List of adverts by user ID",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.MyPreviewAdvertCard"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve adverts by user ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/adverts/saved": {
            "get": {
                "description": "Fetch a list of adverts saved by the specified user ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adverts"
                ],
                "summary": "Retrieve adverts by user ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of adverts saved by user",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PreviewAdvertCard"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve adverts by user ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.Advert": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "has_delivery": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "saves_number": {
                    "type": "integer"
                },
                "seller_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/dto.AdvertStatus"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "views_number": {
                    "type": "integer"
                }
            }
        },
...
        "dto.AdvertRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "has_delivery": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/dto.AdvertStatus"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.MyPreviewAdvertCard": {
            "type": "object",
            "properties": {
                "preview": {
                    "$ref": "#/definitions/dto.PreviewAdvert"
                },
                "saves_number": {
                    "type": "integer"
                },
                "views_number": {
                    "type": "integer"
                }
            }
        },
...
        "dto.PreviewAdvertCard": {
            "type": "object",
            "properties": {
                "is_saved": {
                    "type": "boolean"
                },
                "is_viewed": {
                    "type": "boolean"
                },
                "preview": {
                    "$ref": "#/definitions/dto.PreviewAdvert"
                }
            }
        },
...
        "utils.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}